# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443/
  description: Generated server url
paths:
...
  /api/messages/:
    post:
      tags:
      - message-rest-controller
      operationId: postMessageSecuriced
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Message'
...
  /api/messages/{id}:
    get:
      tags:
      - message-rest-controller
      operationId: getMessage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Message'
  /api/messages/{id1}/{id2}:
    get:
      tags:
      - message-rest-controller
      operationId: getMessages
      parameters:
      - name: id1
        in: path
        required: true
        schema:
          type: string
      - name: id2
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
components:
  schemas:
...
    MessageDTO:
      type: object
      properties:
        content:
          type: string
        sender:
          type: string
        receiver:
          type: string
    Message:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        sendDate:
          type: string
          format: date-time
...